refactor: protected siren and tooling
