+++ cypress/mocks/handlers/grist.ts
@@ -1,0 +1,5 @@
+import { HttpResponse, HttpResponseResolver } from 'msw';
+
+export const gristHandler: HttpResponseResolver = ({ request }) => {
+  return HttpResponse.json([]);
+};
