+++ models/protected-siren/index.ts
@@ -1,6 +1,6 @@
 import { DataStore } from '#clients/data-store';
+import { readFromGrist } from '#clients/external-tooling/grist';
 import { InternalError } from '#models/exceptions';
-import { readFromGrist } from '#utils/integrations/grist';
 import logErrorInSentry from '#utils/sentry';
 import {
   extractSirenFromSiret,
